3.4 Moved accept/decline join requests to TeamDetailActivityTest [Lunn]
Join requests are polled when the team activity is loaded; starting in a different activity is required to test (at the current time)
